Update workspace configurations and improve interrupt handling
Improvements have also been made in the interrupt handling in interrupts.c. This includes adding new functions to increase and decrease interrupt space as well as a signal handler function for SIGUSR1 signals. These changes were necessary for the implementation of better and more efficient interrupt handling in the code.

In the main.c file, casting is included when calling the load_flash function to prevent unexpected behavior specifically when the return value is used for calculations.